// src/app/api/model/getFixedDepositProductsProductIdResponse.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements. See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership. The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GetFixedDepositProductsInterestCalculationType } from './getFixedDepositProductsInterestCalculationType';
import { GetFixedDepositProductsProductIdPreClosurePenalInterestOnType } from './getFixedDepositProductsProductIdPreClosurePenalInterestOnType';
import { GetFixedDepositProductsProductIdInterestCompoundingPeriodType } from './getFixedDepositProductsProductIdInterestCompoundingPeriodType';
import { GetFixedDepositProductsProductIdMaxDepositTermType } from './getFixedDepositProductsProductIdMaxDepositTermType';
import { GetFixedDepositProductsProductIdAccountingMappings } from './getFixedDepositProductsProductIdAccountingMappings';
import { GetFixedDepositProductsProductIdFeeToIncomeAccountMappings } from './getFixedDepositProductsProductIdFeeToIncomeAccountMappings';
import { GetFixedDepositProductsInterestCalculationDaysInYearType } from './getFixedDepositProductsInterestCalculationDaysInYearType';
import { GetFixedDepositProductsProductIdActiveChart } from './getFixedDepositProductsProductIdActiveChart';
import { GetFixedDepositProductsProductIdPenaltyToIncomeAccountMappings } from './getFixedDepositProductsProductIdPenaltyToIncomeAccountMappings';
import { GetFixedDepositProductsProductIdMinDepositTermType } from './getFixedDepositProductsProductIdMinDepositTermType';
import { GetFixedDepositProductsInterestPostingPeriodType } from './getFixedDepositProductsInterestPostingPeriodType';
import { GetFixedDepositProductsProductIdCurrency } from './getFixedDepositProductsProductIdCurrency';


/**
 * GetFixedDepositProductsProductIdResponse
 */
export interface GetFixedDepositProductsProductIdResponse { 
    accountingMappings?: GetFixedDepositProductsProductIdAccountingMappings;
    activeChart?: GetFixedDepositProductsProductIdActiveChart;
    currency?: GetFixedDepositProductsProductIdCurrency;
    description?: string;
    feeToIncomeAccountMappings?: Set<GetFixedDepositProductsProductIdFeeToIncomeAccountMappings>;
    id?: number;
    interestCalculationDaysInYearType?: GetFixedDepositProductsInterestCalculationDaysInYearType;
    interestCalculationType?: GetFixedDepositProductsInterestCalculationType;
    interestCompoundingPeriodType?: GetFixedDepositProductsProductIdInterestCompoundingPeriodType;
    interestPostingPeriodType?: GetFixedDepositProductsInterestPostingPeriodType;
    maxDepositTerm?: number;
    maxDepositTermType?: GetFixedDepositProductsProductIdMaxDepositTermType;
    minDepositTerm?: number;
    minDepositTermType?: GetFixedDepositProductsProductIdMinDepositTermType;
    name?: string;
    penaltyToIncomeAccountMappings?: Set<GetFixedDepositProductsProductIdPenaltyToIncomeAccountMappings>;
    preClosurePenalApplicable?: boolean;
    preClosurePenalInterest?: number;
    preClosurePenalInterestOnType?: GetFixedDepositProductsProductIdPreClosurePenalInterestOnType;
    shortName?: string;
}